+++ lib/src/commands/create_app/steps/cleanup/remove_feature_tags.dart
@@ -10,7 +10,8 @@ Future<void> removeFeatureTags() async {
       'LogoColorScheme',
       'Subscriptions',
       'Appwrite',
-      'Pocketbase'
+      'Pocketbase',
+      'Posthog',
     ]) {
       final directory = Directory('lib');