+++ lib/src/commands/create_app/steps/native_updates/copy_ios_files.dart
@@ -1,0 +1,5 @@
+import 'package:io/io.dart';
+
+Future<void> copyIosFiles(String templatePath) async {
+  await copyPath('$templatePath/ios', 'ios');
+}